+++ data_structures/Range_query_cpp/C++/Update_array.cpp
